Improve action ID rendering in Cache tab (#8141)
Bazel attaches a number of special action IDs to cache requests, such as
`prefetcher` when it downloads the inputs for an action that is executed
locally. Such actions were frequently rendered as a broken link that
treated the special action IDs as digests.

This PR ensures that only actual digests are stylized as links and adds
a ` (for local execution)` suffix to prefetcher downloads.