[raft] in RemoveData, treat replica not marked for removal as a valid case (#8158)
When AddReplica fails before the replica descriptor is added to the
range
descriptor, RemoveReplica won't mark this replica for removal since it
was not
on the range descriptor in the first place. 

In this case, we still want to remove the data from raft, but we don't
need to
remove the data from pebble, there is no raft command in StartShard.