Fail webdriver tests if error banner is displayed (#8128)
We generally want to avoid showing error banners to the user. So, we
should fail webdriver tests if the error banner is displayed at any
point during the test.

This is a follow-up from
https://github.com/buildbuddy-io/buildbuddy/pull/8117 - I verified that
the new test logic fails with that commit reverted.

We show a "Logged out!" banner when logging out, so a few tests are
opted out of this assertion currently. Removing that banner requires
some minor auth changes so I split it into a separate PR.