[raft] don't queue zombie tasks when they are already in queue (#8182)
When we are not able to process all zombies before the scan, we don't
want
to add duplicate entries into the zombie. This can make the length of
queue
metric I am adding in the following PR less representive of the number
of
zombies in the system.

Also, if there are any errors happening in removeZombie, add them back
to the
queue.