[raft] don't do extra work on removed zombies (#8214)
When we detect raft zombies, we queries node host with log info which
contains
all replicas that have been started in the history. This means that
zombies that
have been successfully deleted in the past will always be in the list. 

This PR creates a map to track removed zombies; so that we don't look up
for
range descriptors of removed zombies.

Also track the number of zombies deleted successfully.