Adapt handling of special action IDs to Bazel changes (#8243)
In
https://github.com/bazelbuild/bazel/commit/998e7624093422bee06e65965f8a575d05d57c27,
Bazel started to distinguish between files prefetched as inputs to
locally executions and those that are explicitly requested outputs (e.g.
of top-level targets with `--remote_download_toplevel`).

This commit recognizes these two cases and no attaches any special
meaning to the `prefetcher` action ID, which could be misleading as it
was also assigned to action outputs that are never consumed by a local
execution.