OCI registry that can proxy OCI image pulls to remote registries (#8185)
(Most of this code originally written by @tylerwilliams collaborating
with @bduffany.)

This change introduces an OCI registry that will forward image pull
requests to a remote host (as specified in the `X-Forwarded-Host`
header). Eventually I'd like this registry to serve as a read-through
cache for public OCI registries, but I want to see the bytes flow
end-to-end in the dev environment first.

Supported features:
* pulling OCI images from public registries

Coming Soon™️:
* caching images

Not supported:
* pushing images
* accessing registries that require authentication
* `docker login`

N.B.: @tylerwilliams had a `STOPSHIP` in here for passing credentials to
the remote registry. Since authentication is not supported, I removed
any mention of credentials.

---------

Co-authored-by: Tyler Williams <williams.tyler@gmail.com>