Don't panic in kythe (#8271)
I was stupidly returning the value from pebble.Get directly, forgetting
that it's lifetime is only valid until closer.Close() is called. This
meant that in the rare case we got two simultaneous requests, the bytes
would get invalidated before being read and we'd hit a NPE.

The fix is to copy the bytes before returning them.

One small other fix -- I used ClaimsFromContext to get the groupID to
avoid reparsing a jwt.