Reuse vmexec connections in firecracker.go
This will save a tiny bit of time, but that's not the main reason I'm making the change.

Currently, the FirecrackerExecDialDurationUsec metric records the time for 3 dials per action. The first dial is much slower, but the 2nd and 3rd one drive down the average for the metric, so it seems that the par action dial is faster than it really is.