Provide config options to add additional OAuth scopes for OIDC (#8285)
We have a use case where we need to be able to filter out users on login
by LDAP groups. Our OIDC provider is currently unable to do a group
search on LDAP because the OAuth scopes baked into Buildbuddy do not
include `groups`. This PR would allow users to add custom OAuth scopes
if additional permissions are needed.