Delay dialing the VM exec server
The dial doesn't succeed until the VM exec server is healthy, so we might as well give it so more time and do other work in the meantime. I think this will reduce the dial latency by at least 100ms. For actions that take more than 1s to createWorkspaceImage, I think this will nearly eliminate the dial latency.