Add bytes-proxied metrics to BSProxy and CASProxy servers (#8293)
I thought that I would be able to measure this using the prometheus gRPC
metrics, but I couldn't find one tracking bytes. This will require
updating the dashboards because I changed some of the metric names, will
follow-up with that. I tried this out in dev for a couple of hours and
it looks like we're proxying a lot of bytes there!

![Screenshot 2025-02-03 at 12 06
37 PM](https://github.com/user-attachments/assets/9b962d0f-a217-4f4d-ad89-d5ab7388452e)