periodically trim already-finished tasks from front of the executor queue. (#8297)
currently, we don't prune tasks from executor queues until we're
actually taking on new work. when we're under load, this means some
executors can have artificially long "queues" of tasks that have already
been finished by other executors.

this change works around this case by periodically checking if the
frontmost entry in the task queue has already been finished by another
worker (indicated by the record not existing in redis anymore). when
such an entry is found, the worker will continue picking entries off the
front of the queue until it's either empty or finds a task that hasn't
been completed yet. this should help us scale up more efficiently.

as a toy example: if we have an average queue length of 20, scale up
when the average queue length is >5, and all of our tasks take 20
minutes, then doubling the number of running executors will cut queue
length to 10--so we keep scaling up and quadruple (!) the number of
executors we have running, even if we only needed a few extra executors
in order to take on all tasks.