[RB] Handle retries in the CLI rather than server-side (#8195)
Extracted CLI code from
https://github.com/buildbuddy-io/buildbuddy/pull/8172

This PR disables automatic retries when using the remote bazel CLI. The
CLI then implements retry behavior itself so it can detect when a retry
is happening

This fixes a bug where if the remote run fails due to a transient error,
the CLI will report a failure. However the scheduler may retry the run
in the background and it may eventually succeed. When users go to the
invocation link, they'll see that it succeeded despite expecting a
failure

Slack thread (Issue 2):
https://buildbuddy-corp.slack.com/archives/C07GMM2VBLY/p1734595804049229?thread_ts=1734595606.837269&cid=C07GMM2VBLY

Should be merged after
https://github.com/buildbuddy-io/buildbuddy/pull/8172 is rolled out