Test pulling from OCI registry mirror using `docker` in Firecracker VM (#8362)
I'm targeting the `dockerd`-running-in-Firecracker path as the best
starting place for the caching OCI registry mirror. Before rolling out
the OCI registry mirror to dev and prod, I want to make sure that path
works!

This change tests 3 cases:
1. Pulling an image name that includes the registry mirror address.
2. Pulling the vanilla image name.
3. Having the registry mirror respond 404 to requests for the manifest,
forcing `docker` to fall back to Docker Hub.

To build confidence that the tests actually pull from the mirror, I
count the requests that go to the mirror. It would be fantastic if the
OCI image itself could tell you where the bytes come from, but I have
not found a way of doing that. Suggestions welcome!

---------

Co-authored-by: Brandon Duffany <brandon@buildbuddy.io>