support drilling down into executions by target and mnemonic (#8392)
tacking this onto some existing code for string dimensions--would be
nice to be on a different filters ui, but that's way more work.

to reduce clutter a bit + since these filters don't apply outside of
drilldowns, i've made it so that the text inputs in the filters menu
only show up if the user is already filtering (i.e., they've clicked a
drilldown). we can add some links elsewhere in the UI to see this page
with the filter pre-set like we did for executions per-worker on the
executors page.