[SS] Refactor UFFD to prepare for balloon (#8406)
To make my balloon PR easier to review, refactor the UFFD code to
prepare for balloon-related changes. The major change in this PR is that
rather than {read page fault event, resolve page fault event} in a loop,
we will {read all available page fault events, resolve all page faults}
in a loop. This behavior is required to handle balloon events (comments
coming in a future PR) but should be a no-op for now