Extend deadline for networking cleanup (#8434)
Initial attempt at fixing the `network_cleanup_failed` errors. A lot of
these are failing with "deadline exceeded," so try extending the
deadline by a few seconds.

We are also seeing some of the `iptables` commands fail with `signal:
killed` but I'm not sure if this is the same error (because
`CommandContext` is implemented by sending SIGKILL when the context
deadline expires) or if this is an OOM-kill. Will look into this
separately.