Dedupe task reservations (#8424)
While working on a different issue, I noticed locally that while an
executor was eagerly requesting work during idle periods, the queue
length kept going up. This was surprising because I had enqueued a fixed
set of 100 tasks, but the queue length climbed to over 200.

The issue was that although the executor was under the
`excess_capacity_threshold` on CPU and RAM, it was fully utilized on
custom resources (`gpu`), because all of the tasks in the queue were
requesting 100% of `gpu` resources. So the executor kept requesting more
work, and because we don't dedupe tasks, this drove up the queue length.
To address this specific case, we could incorporate custom resources
into the excess capacity threshold.

But I think the queue can still build up with duplicate tasks in other
cases, too. For example, if the executor is stuck at <40% utilization
because the current task is taking a long time, but the next task in the
queue is very large and cannot schedule, then the executor will keep
requesting more work, likely getting a lot of duplicate tasks. The new
queue pruning logic might help with this, but it's not guaranteed,
especially if multiple executors in a pool are in this state.

Because it's a pretty small change, it seems worth just dropping tasks
if they are already in the queue, just so that we never have to wonder
whether duplicate tasks are inflating the queue length.

This change should not cause legitimate executions to be dropped,
because the only time the same task ID should be legitimately executed
twice is during a retry attempt. In the retry case, we explicitly send
more probes, so retries should still continue to work properly.