[SS] Add memory balloon support in FC (#8460)
Disclaimer: I'm seeing some strange behavior where the balloon fails to
inflate on a snapshotted VM if the balloon was not inflated to the same
amount on the original VM. For example, if the balloon expands 5GB on
the original VM, it can expand 5GB on the snapshot. If the balloon only
expands 4GB on the original VM, it stalls around 4GB on the snapshot,
even when you try to inflate it to 5GB. I'll keep investigating and
maybe report a FC bug, but figured I'd send out this code in the
meantime.

Here's a doc on some various considerations I looked into w.r.t.
performance
https://docs.google.com/document/d/194-3U1Zcn_rJuQ90XXiiGrrLkMqkiUGQp9OhuHXBFdY/edit?tab=t.0