Speculative fix for top-level artifact downloads despite BwoB
When a top-level artifact was produced by a BwoB build, Bazel will download it even with `--remote_download_minimal` after the period specified by `--experimental_remote_cache_ttl` has passed (assuming that the Bazel server is still running). This is unexpected as only the metadata should be refreshed (i.e., a find missing blobs request should be sent).

Test whether this was actually the cause of the large cache downloads we were seeing by setting the period to a very high value.