Speculative fix for top-level artifact downloads despite BwoB (#8502)
When a top-level artifact was produced by a BwoB build, Bazel will
download it even with `--remote_download_minimal` after the period
specified by `--experimental_remote_cache_ttl` has passed (assuming that
the Bazel server is still running). This is unexpected as only the
metadata should be refreshed (i.e., a find missing blobs request should
be sent).

While we wait for the Bazel fix
(https://github.com/bazelbuild/bazel/pull/25398), set this to a high
value to work around the issue. This could negatively impact the
handling of remote cache evictions.