OCI registry mirror that passes through HTTP requests to upstream registry (#8468)
The `ociregistry` is now an OCI registry mirror that supports HEAD and
GET requests to retrieve manifests and blobs. It makes requests to an
upstream registry (`index.docker.io` by default). It does nothing else.

The tests come from
[01_pull_test.go](https://github.com/opencontainers/distribution-spec/blob/main/conformance/01_pull_test.go)
in the OCI distribution spec.

`ociregistry_test` previously relied on `oci.Resolve`. That was a
mistake. Testing the HTTP interface directly gives me more confidence
that this registry will work for different clients out in the wild.

TODO:
1. Cache manifests and blobs in the CAS. See #8510.
2. Remove `X-Forwarded-Host` header from `oci.Resolve`. Unnecessary and
not in any spec.