[raft] use exponential backoff in RunNodehostFn (#8542)
https://github.com/buildbuddy-io/buildbuddy-internal/issues/3745

The number of nodehost.SyncPropose is crazily high during rollout:
![Screenshot from 2025-03-03
17-53-45](https://github.com/user-attachments/assets/aa1dafa1-8c33-482a-88dd-2da3e36eaa3f)

Tested in dev:
This is the p99 of two rollouts after the change is pushed to dev:
![Screenshot from 2025-03-03
23-47-04](https://github.com/user-attachments/assets/21843c19-5123-49a9-91e7-8109c4110f15)

This is the p99 of a typical rollout before the change:
![Screenshot from 2025-03-03
23-52-31](https://github.com/user-attachments/assets/4e87617a-e25d-4695-8097-844ae4c2e2a7)

This is the number of nodehost.SyncPropose calls after the change. The
number stays stable during rollout.
![Screenshot from 2025-03-03
23-30-52](https://github.com/user-attachments/assets/9ce0b394-329e-4b56-a82e-7428a2d486af)