Fix race condition in `bazelisk.Run` (#8569)
Fixes a race condition that occurred when passing a buffer to replace
stderr in `bazelisk.Run`. Previously, both the goroutine that redirected
stderr and the `goLog` had access to the (non-threadsafe) underlying
writer and could attempt to write to it in two separate threads. This PR
replaces the bare pointer to the writer being passed to `goLog` with the
`PipeWriter` we put in front of it, preventing the race condition.