add a counter metric for filecache writes (#8576)
so that we can tell how much _more_ disk we're using when we enable
storing treecache data in the filecache.

the current metric is just a bucketed histogram and is kind of a pain to
read. if you'd rather that we just have the one metric, i can make do.