Store OCI image layers and manifests in the CAS (#8510)
This change adds caching to the OCI registry mirror. It depends on
#8468. Please review #8468 first!

When a client makes a GET request _by digest_ for a blob (layer) or
manifest, the registry mirror will
1. see if the payload exists in the CAS and return it if it does;
2. otherwise, fetch the payload from an upstream registry and
3. store payload in the CAS if successfully found in the upstream
registry.

If there is an error reading from or writing to the CAS, the request
will fail.

People typically fetch manifests by tag (i.e., `latest`). The mapping
from tag to OCI image can change. For example, `latest` changes every
time people upload a new version of an image. Therefore I do not cache
OCI image manifests by tag nor do I cache the mapping from tag to
digest.

This means that `docker pull imagename` or `docker pull
imagename:latest` will make at least one GET request to the upstream
registry, which I think counts against the Docker Hub pull rate limit.
It is possible to perform a HEAD request to the upstream registry for
the manifest tag, read the digest from the response, then look up the
manifest by digest in the CAS. That behavior can come in a follow-on
change.