Populate the bazel commands list for the parser dynamically using the flag collection proto (#8573)
Switches from using a manually-populated slice of bazel commands to one
dynamically generated using the information in the flags collection
proto returned by `bazel help flags-as-proto`.

Mocks out the bazel help function in integration tests that make use of
the parser.

Also silences the call to `bazel help flags-as-proto` in the cli. First,
demoted the preceding `Gathering metadata` message from `Print` to
`Debug`, and secondly, captured the error output of the command itself,
which contained messages about the `TEST_TMPDIR` directory and the
`max_idle_secs` option. Instead of capturing the stderr output, we could
use the `--quiet` option, but that requires bazel 8.0.0 or higher, which
is not guaranteed.