Don't make lookup active replicas during server startup (#8594)
If we have called RemoveData on a node, this node won't show up in
GetNodeHostInfo even when we set skipLogInfo = false. This means that if
we successfully removed a replica from the driver or cleaned up
a zombie, it should not appear in GetNodeHostInfo, therefore won't be
started during server startup.

https://github.com/buildbuddy-io/buildbuddy-internal/issues/4508