Move VFS setup inside the workspace package. (#8604)
Don't expose the VFS scratch space outside the VFS implementation.

Prior to this change, the overall flow was as follows:
 - The workspace is mapped to the scratch space
 - The FUSE mount is exposed to the action via a different path
- Files in the scratch space are laid out using the same paths as in the
fuse mount so that the workspace code can use the scratch space directly
(e.g. upload output files)

After this change, all interactions between the executor and the VFS
implementation happen only over the FUSE mount.

The main motivation for this was an edge case that was hard to fix in
the previous model: An action could create a hardlink to a CAS artifact
and there wasn't a good way to represent this hardlink inside the
scratch workspace.

The change has other benefits:
- The "workspace reuse" code will "just work" since it will no longer
need any special handling for the VFS case.
- We will no longer need to maintain the same layout in the scratch
space as in the fuse mount which will allow us to get rid of some gnarly
code (directory renames) and reduce some IO related to that.

The downside is that there is likely a small performance hit with the
workspace code having to go through FUSE to interact with the workspace
contents. This is something we should be able to mitigate by
short-circuiting the interactions between the workspace and the VFS
implementation, but I'm going to punt on that until the initial VFS
version is shipped.