[raft] remove removedZombies map (#8620)
We added this map because in the past we have seen the replica janitor
keep
removing zombies that have been removed when `RemoveData` was done with
best
effort.

Now that we want to enforce `RemoveData` being called as the last step
in zombie
cleanup. This map can potentially hide errors.