[RB] Fix bug when remote commit doesn't exist locally (#8489)
The remote bazel CLI selects a GitHub branch and commit to fetch on the
remote runner.

There's an edge case where (1) the remote branch has a commit that the
local branch does not and (2) the local commit does not exist remotely.
Previously we would use the head commit from the remote branch. This
would cause issues in the CLI, because we'd try to run `git diff`
operations against a commit that didn't exist locally