Allow a portion of machines to skip saving snapshots
This only applies to action executions which were already created from a snapshot. My hypothesis is that saving an updated snapshot doesn't provide future executions any benefit in these cases.

I choose to use the hostname to split the experiment, because I want every executor on a machine to write less to the filecache if it's in the experiment.