Enable anonymous recycling when API key is unset (#8663)
The `--debug_enable_anonymous_runner_recycling` flag is needed to enable
runner recycling for anonymous executions. This adds a bit of
unnecessary configuration complexity for local development, and possibly
also some on-prem scenarios where executors are run without auth. If
`--executor.api_key` is unset, then we're already running as part of an
anonymous auth setup, so anonymous runner recycling should be allowed.