timing profile: add unit to some time series lanes
Some time series lanes, such as "Memory usage" and "Network usage"
lanes can display a wide range of metric values that made it hard for
users to "guess" the unit being recorded.

Override the event name in Bazel profile with our own display name with
a unit attached to make it easier to read these metrics.