Rename `OptionDefinitionSet` to `Parser` (#8684)
We rename `OptionDefinitionSet` to `Parser` in preparation for
instanitiating it as a singleton containing the only set of
`OptionDefinition`s as well as the set of supported bazel commands, and
also because it has the parsing methods as its member methods.

In addition to it being a more descriptive name, it is also much
shorter, which is generally preferable.