[raft] try other replicas with range-not-present error (#8693)
We return range-not-present error when the range is not present in
openRanges.

The intention of this error is if we hit the wrong machine, then we
don't want
to try all replicas. However, openRanges is not appropriate for this
check; it
didn't represent all replicas that should be on this machine; but what
was
currently open; during rollout or shutting down, this is only a subset
of all
the ranges that should be on the machine.