Make one `Parser` with all the `OptionDefinitions` (#8685)
In order to simplify parsing, we make just one set of `OptionDefintion`s
and store them in the `Parser` alongside the set of known bazel
commands. Before, we separated them out by command to cater to
`CommandLineSchema`, which was designed for use with the usage-based
help parsing; now, however, we have access to all the flag definitions,
and they store within them which commands they support, so we can just
check if options support commands after retrieving the
`OptionDefiniton`.

Also did away with the `All` field in the parser, as it was never read
and made modification of the Parser confusing when testing with options
that would need to be deleted.