only send timeline data with ActionResults when it is explicitly requested. (#8710)
we set `statsPollInterval` to 50ms in
`enterprise/server/remote_execution/container/container.go`, and we
track 7 stats, so we end up recording 140 bytes per second of execution
time.

sending this back to the client when it looks up actionresults adds up
quickly--if an action took a minute, we'll send 8k of data! i think this
is having a big impact on data transfer for CI runs, where lots and lots
of cached actions are fetched for slow tests.