Refactor task_leaser for easier testing (#8720)
It's hard to test `priority_task_scheduler` currently, because it
internally creates a "real" task leaser which does a bunch of low-level
interactions with the SchedulerClient. We could in theory fake out the
`SchedulerClient` and implement all of the functionality needed for the
task leaser to think it's talking to a real scheduler, but this feels
very tedious and seems too low-level for what we're trying to actually
test.

This PR adds a TaskLeaser interface, which would let us plug in a fake
task leaser so that we can more easily test the priority queue logic
without worrying about low-level details of task leasing.