Fix action merging handling of `do_not_cache` and `skip_cache_lookup`
`skip_cache_lookup` should not prevent action merging (it currently does), but `do_not_cache` should (it currently doesn't).
The former is used by Bazel when an action result is discovered to be invalid, but that can't have happened while the merged execution is still executing.
The latter powers `--runs_per_test` and thus won't be useful to detect flakes if action merging is used.

# Conflicts:
#	enterprise/server/remote_execution/execution_server/execution_server.go

# Conflicts:
#	enterprise/server/remote_execution/execution_server/execution_server.go