Refactor hit_tracker to prepare for integration with hit_tracker_service (#8753)
This PR makes a couple of interface changes to the hit_tracker:
1. It changes the `HitTrackerFactory` functions to take an Invocation ID
instead of inferring one and moves the inference to the caller.
2. It separates the logic that records per-invocation hit-tracking data
in Redis from the logic that records server-wide hit-tracking data in
Prometheus.

If you peek at `hit_tracker_service.go` in
https://github.com/buildbuddy-io/buildbuddy/pull/8726/files#diff-320c6a3d49ee6dc20c37549e3843da7a456a0fc3e23f81e512f6209c7a474a91
you can see that the first change will be used to pass the invocation ID
from the `hit_tracker_service` request proto into the hit_tracker, and
the second will be used to record per-invocation state in Redis without
recording Prometheus metrics, for cache-hits served by Cache Proxies.

This PR should make no functional change.