Stop storing github token during github login (#8711)
We don't actually need to store the github token to support github
login. Stop writing it, so the github_token column is exclusively used
for github-related features.

We want to continue using our current (read-write) app for login,
because users have already onboarded with it. (This is very confusing,
but that does not mean that users need to grant write access for login.
Github differentiates between "authenticating" against an app and
"installing" it. In our login flow, we only have users "authenticate"
against the app, which only requests access to the user's email.)

Now, there may be users who use GitHub login, but want to use our
read-only app for workflows for example. We don't want the two github
tokens to clobber each other. Rather than add a new column to clarify
which app is used for github features, reuse the existing column so it
only holds the app installation-related token.

Also add some clarifying comments and text in our UI.