Fix race condition in `bazelisk.Run` when redirecting both stderr and stdout to the same buffer (#8572)
This fixes a race condition `bazelisk.Run` where if both the out and err
buffers were the same, two pipes pointing to the same buffer were
created with accompanying goroutines, and the buffer could thus be
accessed in two different goroutines. This solves the problem by
checking to see if the buffer is the same between both out and err, and
then creating only one pipe if it is. This was previously sidestepped in
`cli/help/help.go` by passing a `lockingbuffer`, but that requirement is
undocumented, and it seems that removing the possibility of the race
condition is likely better than documenting such a requirement for
future ease of use.