sync buildbuddy toolchain in dev_qa.py (#8773)
Fix issues with dev_qa.py while running on 22.04:

- Use the latest BB toolchain version.
- Ensure our toolchains are registered with Bazel
- Trim some flags that are not relevant to toolchain-platform resolution
- Exclude "local" targets in Gazelle, which may cause incompatibility

On the last point: we are still doing RBE on 20.04 image, so some of the
CC/CGO compilation are incompatible with a 22.04 host. Gazelle has some
`go_bazel_test` targets which are tagged with "local" and "exclusive".
These will use CC/CGO deps, which are compiled remotely, to link the
final binary locally, which can cause issues. Excluding them for now
since the purpose here is to verify RBE of dev deployment.