Reinstate "Refactor github logic to prepare for supporting 2 github apps" (#8788)
The original PR is in the first commit.

In the original PR, there was some logic to fetch the correct github app
based on the user context. This broke:
* reporting GitHub statuses in response to webhook events (the context
is not authenticated)
* remote bazel and certain workflow requests if you authenticate using a
group-level API key

In this PR, we use the repo URL to fetch the correct github app instead
(@bduffany suggested we do this originally :shamecube:). I also added
additional validation that one GitHub owner can't install 2 diffrent BB
apps under different group IDs