[raft] remove registry.Add call when we add non-voting replicas (#8796)
https://github.com/buildbuddy-io/buildbuddy-internal/issues/4573

When we start the shard in AddReplica, dragonboat will call registry.Add
on our
behalf so we don't need to call it. Sometimes, a non-voter can be a
zombie, and the shard is never started. If we don't call registry.Add
here, we don't need to clean up.

The node info is either added during preload Registry; or through
RegistryPush
gossip requests.