[raft] implement registry.Remove/RemoveShard (#8795)
https://github.com/buildbuddy-io/buildbuddy-internal/issues/4573

dragonboat calls Remove/RemoveShard when it applies RemoveNode
membership change.

I changed the nodeTargets to a regular map, added a new map to keep
track of all replicas for a single shard. These two maps are protected
by a mutex since they need to be in sync.

Also fix the test. Previously in testutil, we share a single registry
among different stores created StoreFactory and this can cause test to
fail now we are removing entries from registries.