Allow statusz widgets to modify running state and enable pausing executor (#8799)
Siggi talked me out of some iframe shenanigans. There might be some htmx
thingy to do here but I just did the old way. If a statusz section
implements this interface:
```
type StatusUpdater interface {
        // Sections that implement this interface may update the service's
        // current running state in some way.
        UpdateStatusz(w http.ResponseWriter, r *http.Request)
}
```
then it may handle post requests for itself and update state, which
allows putting buttons etc etc in the statusz component that update
things.

In this particular case, we add a button that allows pausing an executor
without failing the health check or stopping it completely. This can be
useful for debugging.

---------

Co-authored-by: Brandon Duffany <brandon@buildbuddy.io>