[FC] Add experiment to skip resaving snapshots
This only applies to action executions which were already created from a snapshot. My hypothesis is that saving an updated snapshot doesn't provide future executions any benefit in these cases.

https://github.com/buildbuddy-io/buildbuddy-internal/pull/4717 is an example for configuring the experiment.

Triggering the experiment by either request metadata or executor hostname makes it more complicated.

Each task is enqueued on multiple executors, and we can't add the platform property to the task until we know which executor will dequeue it, because we only have 1 copy of the task on the scheduler side. This means we have to wait until Scheduler.Lease to set the platform property to triggering on executor hostname.

The executor doesn't have the task (and therefore the platform properties) until Lease returns. It also doesn't have Bazel's RequestMetadata until then, which means Scheduler.Lease doesn't get the metadata either, so we need some hacks to add the metadata values to the experiment context.