Add httpclient that counts outgoing requests; use for OCI image requests & fetch_server (#8707)
This change introduces the `httpclient` package, which can create a
`http.Client` that counts HTTP requests and restricts access to private
IPs.

These packages now use this client:
* fetch_server
* oci
* ociregistry

There are plenty of call sites that do not use this client. I'm
introducing it so I can count the number of requests we're making to
Docker Hub. With the stricter Docker Hub rate limits coming into effect
on April 1st, it'll become crucial to see how many HEAD and GET requests
we're sending to Docker Hub each hour (10 pulls per hour, where a pull =
manifest GET + layer GETs for an image).

~I'll adapt this branch to work with #8697 once it lands.~